Refactored into several classes/files, code cleanup
